# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  responses.LibraryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_public:
        type: boolean
      name:
        type: string
      owner_id:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  responses.MemberResponse:
    properties:
      display_name:
        type: string
      email:
        type: string
      invited_by:
        type: string
      joined_at:
        type: string
      role:
        type: string
      role_id:
        type: string
      tags:
        items:
          $ref: '#/definitions/responses.TagRef'
        type: array
      user_id:
        type: string
      username:
        type: string
    type: object
...
  responses.TagRef:
    properties:
      color:
        type: string
      created_at:
        type: string
      id:
        type: string
      library_id:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Librarium
    url: https://github.com/fireball1725/librarium
  description: Self-hosted library tracking API. All protected endpoints require a
    Bearer JWT in the Authorization header.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  termsOfService: http://localhost/
  title: Librarium API
  version: 26.0.0-dev
paths:
...
  /libraries:
    get:
      description: Returns all libraries the authenticated user has access to.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.LibraryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List libraries
      tags:
      - libraries
    post:
      consumes:
      - application/json
      description: Creates a new library owned by the authenticated user.
      parameters:
      - description: Library details
        in: body
        name: body
        required: true
        schema:
          properties:
            description:
              type: string
            is_public:
              type: boolean
            name:
              type: string
            slug:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.LibraryResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a library
      tags:
      - libraries
  /libraries/{library_id}:
    delete:
      description: Permanently deletes a library and all its contents.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a library
      tags:
      - libraries
    get:
      description: Returns details for a specific library.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.LibraryResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a library
      tags:
      - libraries
    put:
      consumes:
      - application/json
      description: Updates name, description, or visibility of a library.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Updated fields
        in: body
        name: body
        required: true
        schema:
          properties:
            description:
              type: string
            is_public:
              type: boolean
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.LibraryResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a library
      tags:
      - libraries
...
  /libraries/{library_id}/members:
    get:
      description: Returns all members of a library with their roles.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Filter by username/display name
        in: query
        name: search
        type: string
      - description: Filter by tag name
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.MemberResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List library members
      tags:
      - libraries
    post:
      consumes:
      - application/json
      description: Adds an existing user to the library with the specified role.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Member details
        in: body
        name: body
        required: true
        schema:
          properties:
            role:
              type: string
            user_id:
              type: string
          type: object
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a member to a library
      tags:
      - libraries
  /libraries/{library_id}/members/{user_id}:
    delete:
      description: Removes a user's membership from a library. Cannot remove the owner.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove a member from a library
      tags:
      - libraries
    patch:
      consumes:
      - application/json
      description: Changes the role of an existing library member.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      - description: New role
        in: body
        name: body
        required: true
        schema:
          properties:
            role:
              type: string
          type: object
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a member's role
      tags:
      - libraries
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"